feat(wip): flag type select/render
